Version-0.4.1 Progress

+++ src/events/spawn_animated_sprite_event.rs
@@ -4,7 +4,7 @@ use super::spawn_sprite_event::SpawnSpriteEvent;
 
 #[derive(Event)]
 pub struct SpawnAnimatedSpriteEvent {
-    pub sprite_tile_size: f32,
+    pub sprite_tile_size: u32,
     pub frame_timing: f32,
     pub frame_count: usize,
     pub spawn_sprite_event: SpawnSpriteEvent,
